third_party/blink/web_tests/fast/css/font-face-family-setting.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
@font-face {
  font-family: "Foo Bar";
}
</style>
<script>
rule = document.styleSheets[0].cssRules[0];

test(function() {
  assert_equals(rule.style.getPropertyValue('font-family'), 'Foo Bar');
  rule.style.setProperty('font-family', rule.style.getPropertyValue('font-family'));
  assert_equals(rule.style.getPropertyValue('font-family'), 'Foo Bar');
  rule.style.fontFamily = rule.style.fontFamily;
  assert_equals(rule.style.getPropertyValue('font-family'), 'Foo Bar');
  rule.style['font-family'] = rule.style['font-family'];
  assert_equals(rule.style.getPropertyValue('font-family'), 'Foo Bar');
}, 'Should be parsed using font-face descriptor parsing rules and accepted.');

test(function() {
  rule.style.setProperty('font-family', "'Test', serif");
  assert_equals(rule.style.getPropertyValue('font-family'), 'Foo Bar');
  rule.style.setProperty('font-family', "Verdana, Arial, Helvetica");
  assert_equals(rule.style.getPropertyValue('font-family'), 'Foo Bar');
}, 'Should be parsed using font-face descriptor parsing rules and rejected.');

</script>
